modularized and integrated brainstorming and journaling functionalities into commands
